video_duplicate_finder/requirements.txt
opencv-python
imagehash
Pillow
